fix(metrics): upcast prediction tensor to float32 before NumPy under bf16 AMP
_maybe_denormalize_tensor did tensor.detach().cpu().numpy(); under bf16 AMP the
prediction tensor is bfloat16 at the metric/denormalize path (autocast wraps the
whole _step_core, metrics included), and NumPy has no bf16 dtype -> "TypeError:
Got unsupported ScalarType BFloat16", crashing the first train epoch. Add .float()
(lossless bf16->f32, no-op for f32). Only NumPy boundary needed it; the rest of
the metric math stays in torch. CPU-unit-verified with a bf16 input.

Also refresh CLAUDE.md: correct AMP autocast scope (whole _step_core incl.
metrics), record the env upgrade to transformers 5.9.0 + numpy 2.4.4 +
scikit-learn 1.8.0 (resolves the earlier numpy-1.x ABI mismatch), and note
PatchTSMixer still has supports_gradient_checkpointing=False in 5.9.0.

+++ src/processors/metrics_utils.py
@@ -562,7 +562,11 @@ def _maybe_denormalize_tensor(
         )
 
     original_device = tensor.device
-    tensor_np = tensor.detach().cpu().numpy()
+    # Upcast to float32 before NumPy: NumPy has no bfloat16 dtype, so a bf16
+    # tensor (e.g. produced under bf16 AMP autocast) would raise
+    # "Got unsupported ScalarType BFloat16". float() is lossless for bf16->f32
+    # and a no-op for tensors already in float32.
+    tensor_np = tensor.detach().cpu().float().numpy()
     tensor_np = global_min_max_norm(
         tensor_np,
         global_min_max={"min": global_min, "max": global_max},
